+++ docs/make.jl
@@ -1,11 +1,16 @@
-using Documenter
+# SPDX-License-Identifier: MIT
 
-push!(LOAD_PATH, "../")
+push!(LOAD_PATH, "../src")
+
+using Documenter
 using MarkovModels
 
+DocMeta.setdocmeta!(MarkovModels, :DocTestSetup,
+                    :(using MarkovModels), recursive = true)
+
 makedocs(
-    sitename = "MarkovModels",
-    format = Documenter.HTML(prettyurls = get(ENV, "CI", nothing) == "true"),
+    sitename = "MarkovModels.jl",
+    modules = [MarkovModels],
     pages = [
         "Home" => "index.md",
         "Manual" => Any[
@@ -16,5 +21,5 @@ makedocs(
 )
 
 deploydocs(
-    repo = "github.com/BUTSpeechFIT/MarkovModels.git",
+    repo = "github.com/lucasondel/MarkovModels.jl.git",
 )